requirements.txt
fastapi>=0.111.0
langchain-text-splitters>=0.2.1
openai>=1.33
pandas>=2.2.2
pinecone-client[grpc]>=4.1.1
pydantic>=2.6.1
pydantic-settings>=2.3.1
seaborn>=0.13.2
tqdm>=4.66.4
typer>=0.9.0
